requirements.txt
flask
flask-cors
assemblyai
requests
instaloader
python-dotenv
openai